Lazy load job logs page
